Fixed incorrect data type of trans_slot_id
In xl_zheap_unused, data type of trans_slot_id was uint8 that was
wrong because slot can be greater than 256.   If slot is greater than
256, then also we were wal logging it as uint8.  Due to this, we
were getting inconsistant page found.

Patch by Mahendra Singh Thalor, reviewed by Dilip Kumar

+++ src/include/access/zheapam_xlog.h
@@ -278,7 +278,7 @@ typedef struct xl_zheap_unused
 
 	TransactionId latestRemovedXid;
 	uint16		nunused;
-	uint8		trans_slot_id;
+	uint16		trans_slot_id;
 	uint8		flags;
 	/* OFFSET NUMBERS are in the block reference 0 */
 } xl_zheap_unused;
